Goals as data - Phase 2
Extract goal data and options out into functions.  We still end up
with a single goals global with this approach due to the
addSpringSupport function requiring most goals.  I have an idea on how
to work around that but want to commit this progress.

Remove goals that had no-op executors.