Move goal data back to index, move support out
Move construction of the goal data back to the index.

Move configuration of support for various technologies to separate
function and file.

I am not convinced we have really gotten rid of the global goals.  We
have just wrapped them in a function and only use them directly from
one place.